インストラーの強化。プロセスの起動確認を厳密にした

+++ lib/installer/daemon_ctrl.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 )
 
 type DaemonCtrl struct{}
@@ -57,7 +58,14 @@ func (d *DaemonCtrl) GeneratePlistStr() string {
 `, d.GetDaemonName(), binPath, logPath, logPath)
 }
 
-func (d *DaemonCtrl) CreatePListFile() error {
+func (d *DaemonCtrl) CreatePListFile(confirmOverwite bool) error {
+	if common.FileExists(d.GetPListPath()) {
+		fmt.Printf("常駐プロセス(LaunchAgents)ファイルが既に存在します: %s", d.GetPListPath())
+		if confirmOverwite && PrintAndScanStdInput("上書きしますか？ (y/n) > ") != "y" {
+			fmt.Println("常駐プロセス(LaunchAgents)ファイルの上書きを中止しました")
+			return nil
+		}
+	}
 	err := os.WriteFile(d.GetPListPath(), []byte(d.GeneratePlistStr()), 0644)
 	if err != nil {
 		return fmt.Errorf("常駐プロセス(LaunchAgents)ファイルの作成に失敗しました. エラー: %v", err)
@@ -84,13 +92,6 @@ func (d *DaemonCtrl) StartDaemon() error {
 	if err != nil {
 		return fmt.Errorf("常駐プロセス(LaunchAgents)の開始に失敗しました. エラー: %v", err)
 	}
-	isRunning, err := d.IsDaemonRunning()
-	if err != nil {
-		return fmt.Errorf("常駐プロセス(LaunchAgents)の起動確認に失敗しました. エラー: %v", err)
-	}
-	if !isRunning {
-		return fmt.Errorf("常駐プロセス(LaunchAgents)のコマンドは成功しましたが、プロセスは起動できませんでした")
-	}
 	fmt.Println("常駐プロセス(LaunchAgents)を開始しました")
 	return nil
 
@@ -110,18 +111,36 @@ func (d *DaemonCtrl) StopDaemon() error {
 }
 
 func (d *DaemonCtrl) IsDaemonRunning() (bool, error) {
-	stdout, err := exec.Command("launchctl", "list", d.GetDaemonName()).Output()
+	stdout, err := exec.Command("launchctl", "list").Output()
 	if err != nil {
-		fmt.Println("常駐プロセス(LaunchAgents)が起動していません")
-		return false, nil
+		return false, fmt.Errorf("launchctl list のコマンドの実行に失敗しました: %v", err)
 	}
-	//stdoutの一文字目が「-」であれば、常駐プロセスが起動していない
-	pid := stdout[0]
-	if pid == '-' {
-		fmt.Println("常駐プロセス(LaunchAgents)が起動していません")
-		return false, nil
-	} else {
-		fmt.Printf("常駐プロセス(LaunchAgents)が起動しています。PID: %v\n", pid)
-		return true, nil
+	stdoutStr := string(stdout)
+	// \nで分割して、行ごとに処理
+	lines := strings.Split(stdoutStr, "\n")
+	a := d.GetDaemonName()
+	fmt.Println(a)
+	for _, line := range lines {
+		if strings.Contains(line, d.GetDaemonName()) {
+			// 正常の場合、lineは以下のようになっている
+			// 3113    -       com.github.fetaro.gcal_run
+			// 異常の場合、lineは以下のようになっている
+			// -       78      com.github.fetaro.gcal_run
+			pid := strings.Split(line, "\t")[0]
+			errorStr := strings.Split(line, "\t")[1]
+			if errorStr == "78" {
+				fmt.Println("常駐プロセス(LaunchAgents)が起動していません。おそらくバイナリファイルに実行権限がありません。errorNo=", errorStr)
+				return false, nil
+			} else if pid == "-" {
+				fmt.Println("常駐プロセス(LaunchAgents)が起動していません。errorNo=", errorStr)
+				return false, nil
+			} else {
+				fmt.Printf("常駐プロセス(LaunchAgents)が起動しています。PID: %s\n", pid)
+				return true, nil
+			}
+		}
 	}
+	fmt.Println("常駐プロセス(LaunchAgents)が登録されておらず、起動していません")
+	return false, nil
+
 }
